исправлена часть структуры приложения, удалены дубликаты, созданы новые компоненты
