+++ task_manager/src/App.js
@@ -3,7 +3,7 @@ import React from "react";
 import MainPage from "./pages/MainPage";
 import RegistrationPage from "./pages/RegistrationPage";
 
-import {Route, Switch} from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 import LoginPage from "./pages/LoginPage";
 import { connect } from "react-redux";
 import TasksInProjectPage from "./pages/TasksInProjectPage";
@@ -21,18 +21,18 @@ function App(props) {
   return (
     <div className="App">
       <Switch>
-      <Route path="/projects" exact render={() => <ProjectPage />} />
-      <Route
-        path={`/projects/:id`}
-        exact
-        render={() => <TasksInProjectPage props={props.state.tasks} />}
-      />
-      <Route path="/" exact render={() => <MainPage />} />
-      <Route path="/login" exact component={LoginPage} />
-      <Route path="/registration" exact component={RegistrationPage} />
-      <Route path="/settings" exact component={SettingsPage} />
-      <Route  component={Notfound} />
-        </Switch>
+        <Route path="/projects" exact render={() => <ProjectPage />} />
+        <Route
+          path={`/projects/:id`}
+          exact
+          render={() => <TasksInProjectPage props={props.state.tasks} />}
+        />
+        <Route path="/" exact render={() => <MainPage />} />
+        <Route path="/login" exact component={LoginPage} />
+        <Route path="/registration" exact component={RegistrationPage} />
+        <Route path="/settings" exact component={SettingsPage} />
+        <Route component={Notfound} />
+      </Switch>
     </div>
   );
 }
