+++ task_manager/src/component/GetUsers/index.js
@@ -1,0 +1,1 @@
+export { default } from "./GetUsers";
