+++ task_manager/src/component/ListProjects/ListProjects.scss
@@ -1,0 +1,54 @@
+@import "src/color";
+.status {
+  &-active {
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 14px;
+    color: $Confirm;
+  }
+  &-stop {
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 14px;
+    color: $Error;
+  }
+  &-complete {
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 14px;
+    color: $Grey;
+  }
+}
+
+.projects-list {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  &__item {
+    justify-content: space-between;
+    text-decoration: none;
+    color: black;
+    min-height: 110px;
+    background-color: #ffffff;
+    margin-bottom: 24px;
+    box-shadow: 0 0 15px rgba(2, 57, 164, 0.05);
+    border-radius: 10px;
+    display: flex;
+    flex-direction: column;
+    box-sizing: border-box;
+    flex-grow: 1;
+    flex-shrink: 1;
+    flex-basis: 0;
+    margin-right: 15px;
+    padding: 16px 24px;
+
+    &__title {
+      display: flex;
+      align-items: center;
+      font-weight: normal;
+      font-size: 14px;
+      line-height: 160%;
+    }
+  }
+}
