+++ task_manager/src/component/ListProjects/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ListProjects";
