+++ task_manager/src/component/MenuAndHeader/MenuAndHeader.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Header from "../Header";
+import Menu from "../Menu";
+
+function MenuAndHeader() {
+  return (
+   <div>
+<Header/>
+<Menu/>
+   </div>
+  );
+}
+
+export default MenuAndHeader;
