+++ task_manager/src/component/MenuAndHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MenuAndHeader";
