+++ task_manager/src/component/ModalAddTask/ModalAddTask.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import "./ModalAddTask.scss";
 
 import Button from "../Button";
@@ -10,77 +10,13 @@ import {
 import { connect } from "react-redux";
 import { ReactComponent as Closed } from "../../svg/closed.svg";
 import Users from "../Users/Users";
+import { useInput } from "../Validation";
 
 let mapStateToProps = state => {
   return {
     state: state.main,
     user: state.users.profile,
     users: state.users.users
-  };
-};
-
-const useInput = (initialValue, validations) => {
-  const [value, setValue] = useState(initialValue);
-  const [isDirty, setDirty] = useState(false);
-  const valid = useValidation(value, validations);
-  const onChange = e => {
-    setValue(e.target.value);
-  };
-  const onBlur = () => {
-    setDirty(true);
-  };
-  return {
-    value,
-    onChange,
-    onBlur,
-    ...valid,
-    isDirty
-  };
-};
-
-const useValidation = (value, validations) => {
-  const [inputValid, setInputValid] = useState(false);
-  const [isEmpty, setEmpty] = useState(true);
-  const [minLengthError, setMinLengthError] = useState(false);
-  const [maxLengthError, setMaxLengthError] = useState(false);
-  const [emailError, setEmailError] = useState(false);
-  useEffect(() => {
-    for (const validation in validations) {
-      switch (validation) {
-        case "minLength":
-          value.length < validations[validation]
-            ? setMinLengthError(true)
-            : setMinLengthError(false);
-          break;
-        case "maxLength":
-          value.length > validations[validation]
-            ? setMaxLengthError(true)
-            : setMaxLengthError(false);
-          break;
-        case "isEmpty":
-          value ? setEmpty(false) : setEmpty(true);
-          break;
-        case "isEmail":
-          const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-          return re.test(String(value).toLowerCase())
-            ? setEmailError(false)
-            : setEmailError(true);
-          break;
-      }
-    }
-  }, [value]);
-  useEffect(() => {
-    if (isEmpty || maxLengthError || minLengthError || emailError) {
-      setInputValid(false);
-    } else {
-      setInputValid(true);
-    }
-  }, [isEmpty, maxLengthError, minLengthError, emailError]);
-  return {
-    isEmpty,
-    minLengthError,
-    maxLengthError,
-    inputValid
   };
 };
