+++ task_manager/src/component/Validation/Validation.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "./Validation.scss";
+
+function Validation() {
+
+  return <></>
+}
+
+export default Validation;
