+++ task_manager/src/component/Validation/index.js
@@ -1,0 +1,73 @@
+import {useEffect, useState} from "react";
+
+export const  useInput = (initialValue, validations) => {
+  const [value, setValue] = useState(initialValue);
+  const [isDirty, setDirty] = useState(false);
+  const valid = useValidation(value, validations);
+  const onChange = e => {
+    setValue(e.target.value);
+  };
+  const onBlur = () => {
+    setDirty(true);
+  };
+  const deleteValue = () => {
+    setValue("");
+    setDirty(false);
+  };
+  return {
+    value,
+    onChange,
+    onBlur,
+    ...valid,
+    isDirty,
+    deleteValue
+  };
+};
+
+export const useValidation = (value, validations) => {
+  const [inputValid, setInputValid] = useState(false);
+  const [isEmpty, setEmpty] = useState(true);
+  const [minLengthError, setMinLengthError] = useState(false);
+  const [maxLengthError, setMaxLengthError] = useState(false);
+  const [emailError, setEmailError] = useState(false);
+  useEffect(() => {
+    for (const validation in validations) {
+      switch (validation) {
+        case "minLength":
+          value.length < validations[validation]
+            ? setMinLengthError(true)
+            : setMinLengthError(false);
+          break;
+        case "maxLength":
+          value.length > validations[validation]
+            ? setMaxLengthError(true)
+            : setMaxLengthError(false);
+          break;
+        case "isEmpty":
+          value ? setEmpty(false) : setEmpty(true);
+          break;
+        case "isEmail":
+          const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+          return re.test(String(value).toLowerCase())
+            ? setEmailError(false)
+            : setEmailError(true);
+          break;
+        default:break;
+      }
+    }
+  }, [value]);
+  useEffect(() => {
+    if (isEmpty || maxLengthError || minLengthError || emailError) {
+      setInputValid(false);
+    } else {
+      setInputValid(true);
+    }
+  }, [isEmpty, maxLengthError, minLengthError, emailError]);
+  return {
+    emailError,
+    isEmpty,
+    minLengthError,
+    maxLengthError,
+    inputValid
+  };
+};
