+++ task_manager/src/pages/MainPage/MainPage.jsx
@@ -1,17 +1,15 @@
 import React, { useState } from "react";
 import "./MainPage.scss";
-import Menu from "../../component/Menu";
-import { NavLink, Redirect } from "react-router-dom";
-import Header from "../../component/Header";
+import { Redirect } from "react-router-dom";
 import { connect } from "react-redux";
-import Users from "../../component/Users";
 import Button from "../../component/Button";
 import ModalAddProject from "../../component/ModalAddProject";
+import ListProjects from "../../component/ListProjects";
+import MenuAndHeader from "../../component/MenuAndHeader";
 
 let mapStateToProps = state => {
   return {
     state: state.main,
-    users: state.users.users,
     user: state.users.profile.id
   };
 };
@@ -19,12 +17,10 @@ let mapStateToProps = state => {
 function MainPage(props) {
   const [visibleAddProject, setVisibleAddProject] = useState(false);
   if (!props.user) return <Redirect to={"/login"} />;
+
   return (
     <div className="main-page">
-      <div>
-        <Header />
-        <Menu />
-      </div>
+      <MenuAndHeader />
 
       <div>
         <div className="main-page__content">
@@ -34,48 +30,11 @@ function MainPage(props) {
               <Button text="Добавить проект" type="add-task" color="blue" />
             </div>
           </div>
-          <div className="main-page__content__projects">
-            {props.state.projects.slice(0, 4).map(project => (
-              <NavLink
-                to={`/projects/${project.id}`}
-                className="main-page__content__project"
-                key={project.id}
-              >
-                <div className="main-page__content__project-title">
-                  {project.name}
-                </div>
-                {project.status === 1 && (
-                  <div className="status-active">Активный</div>
-                )}
-                {project.status === 2 && (
-                  <div className="status-stop">Приостановлен</div>
-                )}
-                {project.status === 3 && (
-                  <div className="status-complete">Завершен</div>
-                )}
-                <div className="main-page__content__project-item">
-                  {project.users &&
-                    project.users.map(usersId => (
-                      <div key={`${Math.random()}${usersId}`}>
-                        {props.users.map(
-                          user =>
-                            user.id === usersId && (
-                              <div
-                                key={user.id}
-                                className="main-page__content__project-item-user "
-                              >
-                                <Users
-                                  userIdArray={user.name.substring(0, 1)}
-                                />
-                              </div>
-                            )
-                        )}
-                      </div>
-                    ))}
-                </div>
-              </NavLink>
-            ))}
-          </div>
+          <ListProjects
+            projects={props.state.projects}
+            slice={9}
+            users={props.users}
+          />
         </div>
       </div>
       {visibleAddProject && (
