+++ task_manager/src/pages/NotfoundPage/NotfoundPage.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 
 import "./NotfoundPage.scss";
-import Header from "../../component/Header";
 import {NavLink, Redirect} from "react-router-dom";
 import {connect} from "react-redux";
+import MenuAndHeader from "../../component/MenuAndHeader";
 
 let mapStateToProps = state => {
   return {
@@ -14,7 +14,7 @@ function Notfound(props) {
   if (!props.user) return <Redirect to={"/login"} />;
   return (
     <div className="not-found">
-      <Header />
+      <MenuAndHeader/>
       <div className="not-found-content">
         <div className={"not-found-code"}>404</div>
         <div>Страница не найдена, пожалуйста,</div>
