+++ task_manager/src/pages/SettingsPage/SettingsPage.jsx
@@ -1,14 +1,13 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 import "./SettingsPage.scss";
-import Menu from "../../component/Menu";
-import Header from "../../component/Header";
 import { connect } from "react-redux";
 import Button from "../../component/Button";
 import { updatePassword, updateUser } from "../../redux/profile-reducer";
 import { Redirect } from "react-router-dom";
 import ModalWindow from "../../component/ModalWindow";
-import { minLength } from "../../component/FormAddProject/FormAddProject";
+import { useInput } from "../../component/Validation";
+import MenuAndHeader from "../../component/MenuAndHeader";
 
 let mapStateToProps = state => {
   return {
@@ -16,77 +15,6 @@ let mapStateToProps = state => {
     user: state.users.profile.id,
     updatePasswordMs: state.users.updatePassword,
     updateUsersMs: state.users.updateUsers
-  };
-};
-
-const useInput = (initialValue, validations) => {
-  const [value, setValue] = useState(initialValue);
-  const [isDirty, setDirty] = useState(false);
-  const valid = useValidation(value, validations);
-  const onChange = e => {
-    setValue(e.target.value);
-  };
-  const onBlur = e => {
-    setDirty(true);
-  };
-  const deleteValue = () => {
-    setValue("");
-    setDirty(false);
-  };
-  return {
-    value,
-    onChange,
-    onBlur,
-    ...valid,
-    isDirty,
-    deleteValue
-  };
-};
-
-const useValidation = (value, validations) => {
-  const [inputValid, setInputValid] = useState(false);
-  const [isEmpty, setEmpty] = useState(true);
-  const [minLengthError, setMinLengthError] = useState(false);
-  const [maxLengthError, setMaxLengthError] = useState(false);
-  const [emailError, setEmailError] = useState(false);
-  useEffect(() => {
-    for (const validation in validations) {
-      switch (validation) {
-        case "minLength":
-          value.length < validations[validation]
-            ? setMinLengthError(true)
-            : setMinLengthError(false);
-          break;
-        case "maxLength":
-          value.length > validations[validation]
-            ? setMaxLengthError(true)
-            : setMaxLengthError(false);
-          break;
-        case "isEmpty":
-          value ? setEmpty(false) : setEmpty(true);
-          break;
-        case "isEmail":
-          const re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-          return re.test(String(value).toLowerCase())
-            ? setEmailError(false)
-            : setEmailError(true);
-          break;
-      }
-    }
-  }, [value]);
-  useEffect(() => {
-    if (isEmpty || maxLengthError || minLengthError || emailError) {
-      setInputValid(false);
-    } else {
-      setInputValid(true);
-    }
-  }, [isEmpty, maxLengthError, minLengthError, emailError]);
-  return {
-    emailError,
-    isEmpty,
-    minLengthError,
-    maxLengthError,
-    inputValid
   };
 };
 
@@ -143,10 +71,7 @@ function SettingsPage(props) {
   if (!props.user) return <Redirect to={"/login"} />;
   return (
     <div className="main-page">
-      <div>
-        <Header />
-        <Menu />
-      </div>
+      <MenuAndHeader />
       <div className="main-page__content">
         <div className="1">
           <div className=" settings-page__container">
